# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get users",
        "parameters": [
          {
            "name": "subsidiaryId",
            "required": true,
            "in": "query",
            "description": "The subsidiary id to filter users",
            "example": "f7b9e3b0-4b7b-4b7b-8b7b-4b7b7b7b7b7b",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all users"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a new user"
          },
          "400": {
            "description": "User already exists"
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/{id}": {
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update a user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated succesfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Delete a user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted succesfully"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Observatorio USS",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreatePersonDto": {
        "type": "object",
        "properties": {
          "givenNames": {
            "type": "string",
            "description": "The given names of the person",
            "example": "Jhon"
          },
          "lastName": {
            "type": "string",
            "description": "The last name of the person",
            "example": "Doe"
          },
          "address": {
            "type": "string",
            "description": "The address of the person",
            "example": "Av. Lima 123"
          },
          "email": {
            "type": "string",
            "description": "The email of the person",
            "example": "jhon@gmail.com"
          },
          "phone": {
            "type": "string",
            "description": "The phone of the person",
            "example": "987654321"
          },
          "image": {
            "type": "string",
            "description": "The image of the person",
            "example": "https://example.com/image.jpg",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The type of the person",
            "example": "natural"
          },
          "documentType": {
            "type": "string",
            "description": "The document type of the person",
            "example": "dni"
          },
          "documentNumber": {
            "type": "string",
            "description": "The document number of the person",
            "example": "12345678"
          }
        },
        "required": [
          "givenNames",
          "lastName",
          "address",
          "email",
          "phone",
          "image",
          "type",
          "documentType",
          "documentNumber"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "jhondoe1"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "subsidiaryId": {
            "type": "string",
            "description": "The subsidiary id of the user",
            "example": "f7b1b3b0-1b1b-4b1b-8b1b-1b1b1b1b1b1b"
          },
          "person": {
            "description": "The person data of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreatePersonDto"
              }
            ]
          },
          "roleName": {
            "type": "string",
            "description": "The role of the user",
            "example": "cashier",
            "enum": [
              "super_admin",
              "admin",
              "cashier",
              "warehouse_manager",
              "accountant"
            ]
          }
        },
        "required": [
          "username",
          "password",
          "subsidiaryId",
          "person",
          "roleName"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "The username of the user",
            "example": "jhondoe1"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          },
          "subsidiaryId": {
            "type": "string",
            "description": "The subsidiary id of the user",
            "example": "f7b1b3b0-1b1b-4b1b-8b1b-1b1b1b1b1b1b"
          },
          "person": {
            "description": "The person data of the user",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreatePersonDto"
              }
            ]
          },
          "roleName": {
            "type": "string",
            "description": "The role of the user",
            "example": "cashier",
            "enum": [
              "super_admin",
              "admin",
              "cashier",
              "warehouse_manager",
              "accountant"
            ]
          },
          "isActive": {
            "type": "boolean",
            "description": "The status of the user",
            "example": true
          }
        },
        "required": [
          "isActive"
        ]
      },
...
    }
  }
}